+++ src/OpenTelemetry.Instrumentation.Cassandra/CHANGELOG.md
@@ -5,8 +5,8 @@
 * Updated CassandraCSharpDriver to `3.17.0`.
   ([#2836](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2836))
 
-* Updated OpenTelemetry core component version(s) to `1.13.0`.
-  ([#3158](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3158))
+* Updated OpenTelemetry core component version(s) to `1.13.1`.
+  ([#3218](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3218))
 
 ## 1.0.0-beta.2